gcp_functions/summarize_rfp/requirements.txt
fastapi
uvicorn
pdfminer.six
openai